backend/requirements-v2.txt
flask
flask-cors
flask-socketio
ollama
chromadb
python-dotenv
pytesseract
eventlet
gevent
pandas
pandasai
pandas_access
pymongo
pydantic
llama-index
llama-index-readers-web
gunicorn
markupsafe
werkzeug
translators